+++ packages/core/src/__tests__/data/string-literal-store.raw.tsx
@@ -1,0 +1,8 @@
+import { useStore } from '@builder.io/morpho';
+
+export default function StringLiteralStore() {
+  /* prettier-ignore */
+  const state = useStore({ "foo": 123 });
+
+  return <div>{state.foo}</div>;
+}